# swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "InnoScope API",
		"version": "1.0",
		"description": "InnoScope is DevScope's intern project for idea submission and realization into teams. It started at 2023's Academy."
	},
	"produces": [
		"application/json"
	],
	"schemes": [
		"http"
	],
	"components": {
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
...
		"parameters": {
...
			"ideaId": {
				"name": "ideaId",
				"in": "path",
				"description": "Idea ID",
				"required": true,
				"schema": {
					"type": "integer",
					"minimum": 1
				}
			},
...
			"per_page": {
				"name": "per_page",
				"in": "query",
				"description": "Number of itens per page",
				"schema": {
					"type": "integer",
					"minimum": 1,
					"default": 20
				}
			},
			"curr_page": {
				"name": "curr_page",
				"in": "query",
				"description": "Number of current page",
				"schema": {
					"type": "integer",
					"minimum": 1,
					"default": 1
				}
			},
			"title": {
				"name": "title",
				"in": "query",
				"description": "Title to search",
				"schema": {
					"type": "string"
				}
			},
			"category": {
				"name": "category",
				"in": "query",
				"description": "Number of category to filter",
				"schema": {
					"type": "integer",
					"minimum": 1
				}
			},
			"status": {
				"name": "status",
				"in": "query",
				"description": "Status type to filter",
				"schema": {
					"type": "string",
					"enum": [
						"New",
						"On Voting",
						"Rejected",
						"Approved",
						"Waiting",
						"On Going",
						"Finished"
					]
				}
			},
			"sort": {
				"name": "sort",
				"in": "query",
				"description": "Sort type to order",
				"schema": {
					"type": "string",
					"enum": [
						"Name (A-Z)",
						"Name (Z-A)",
						"Creation (New-Old)",
						"Creation (Old-New)"
					]
				}
			}
		},
		"responses": {
			"unauthorized": {
				"description": "Unauthorized",
				"content": {
					"application/json": {
						"schema": {
							"type": "object"
						},
						"examples": {
							"No token": {
								"value": {
									"success": false,
									"msg": "You need to be authenticated to access to this route."
								}
							},
							"Invalid token": {
								"value": {
									"success": false,
									"msg": "Unauthorized."
								}
							}
						}
					}
				}
			},
			"unexistent_idea": {
				"description": "Unexistent idea id",
				"content": {
					"application/json": {
						"schema": {
							"type": "object"
						},
						"examples": {
							"Unexistent idea id": {
								"value": {
									"error": "Idea not found."
								}
							}
						}
					}
				}
			}
		}
	},
...
	"paths": {
...
		"/ideas": {
			"get": {
				"summary": "Get all ideas in the system",
				"operationId": "getAllCommunityIdeas",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"$ref": "#/components/parameters/per_page"
					},
					{
						"$ref": "#/components/parameters/curr_page"
					},
					{
						"$ref": "#/components/parameters/title"
					},
					{
						"$ref": "#/components/parameters/category"
					},
					{
						"$ref": "#/components/parameters/status"
					},
					{
						"$ref": "#/components/parameters/sort"
					}
				],
				"responses": {
					"200": {
						"description": "List of users",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"List of users": {
										"value": {
											"ideas": [],
											"pagination": {
												"total": 0,
												"per_page": 20,
												"curr_page": 1,
												"prev_page": null,
												"next_page": null,
												"offset": 0,
												"to": 20,
												"last_page": 1
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/unauthorized"
					},
					"422": {
						"description": "Invalid fields",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Invalid per page": {
										"value": {
											"errors": [
												"Please enter a valid per page quantity."
											]
										}
									},
									"Invalid current page": {
										"value": {
											"errors": [
												"Please enter a valid current page number."
											]
										}
									},
									"Invalid title filter": {
										"value": {
											"errors": [
												"Please enter a valid title."
											]
										}
									},
									"Invalid category filter": {
										"value": {
											"errors": [
												"Please enter a valid category."
											]
										}
									},
									"Invalid status filter": {
										"value": {
											"errors": [
												"Please enter a valid status."
											]
										}
									},
									"Invalid sort order": {
										"value": {
											"errors": [
												"Please enter a valid sort type."
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/ideas/{ideaId}": {
			"get": {
				"summary": "Get one idea details",
				"operationId": "getCommunityIdeaDetails",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"$ref": "#/components/parameters/ideaId"
					}
				],
				"responses": {
					"200": {
						"description": "Idea item",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Idea item": {
										"value": {
											"idea": {
												"idea_id": 1,
												"idea_title": "Idea",
												"idea_summary": "Brief description",
												"idea_description": "Big description of idea",
												"idea_complexity": "Easy",
												"idea_durationWeeks": 1,
												"idea_status": "New",
												"categoryId": 1,
												"category": "Name",
												"technologies": [
													{
														"technology_id": 1,
														"technology_name": "TypeScript"
													}
												]
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/unexistent_idea"
					},
					"422": {
						"description": "Invalid fields",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Invalid idea id": {
										"value": {
											"errors": [
												"Please enter a valid idea id."
											]
										}
									}
								}
							}
						}
					}
				}
			},
			"patch": {
				"summary": "Update one idea status to next status available",
				"operationId": "updateIdeaStatus",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"$ref": "#/components/parameters/ideaId"
					}
				],
				"responses": {
					"200": {
						"description": "Idea updated",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Idea updated": {
										"value": {
											"message": "Idea status updated successfully."
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Error on update",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Cannot update": {
										"value": {
											"error": "Idea status cannot be updated."
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/unexistent_idea"
					},
					"422": {
						"description": "Invalid fields",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Invalid idea id": {
										"value": {
											"errors": [
												"Please enter a valid idea id."
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/ideas/{ideaId}/interactions": {
			"get": {
				"summary": "Get votes and comments of idea",
				"operationId": "getIdeaInteractions",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"$ref": "#/components/parameters/ideaId"
					}
				],
				"responses": {
					"200": {
						"description": "Interactions list",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Idea interactions": {
										"value": {
											"voteAvg": 4,
											"comments": [
												{
													"comment": "text",
													"user": "John Doe"
												}
											]
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/unexistent_idea"
					},
					"422": {
						"description": "Invalid fields",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Invalid idea id": {
										"value": {
											"errors": [
												"Please enter a valid idea id."
											]
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Add vote and/or comment to idea",
				"operationId": "addIdeaInteraction",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"$ref": "#/components/parameters/ideaId"
					}
				],
				"requestBody": {
					"description": "comment/vote to add to idea",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object"
							},
							"examples": {
								"Vote": {
									"summary": "Vote to add to idea",
									"value": {
										"vote": 4
									}
								},
								"Comment": {
									"summary": "Comment to add to idea",
									"value": {
										"comment": "text"
									}
								},
								"Vote+Comment": {
									"summary": "Vote and comment to add to idea",
									"value": {
										"vote": 4,
										"comment": "text"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Interaction creation",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Interaction created/updated": {
										"value": {
											"message": "Interaction added or updated."
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Interaction creation error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Idea not open for interaction": {
										"value": {
											"message": "Idea is not open for interaction."
										}
									},
									"Idea's author interacting": {
										"value": {
											"message": "You can't interact with your idea."
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/unexistent_idea"
					},
					"422": {
						"description": "Invalid fields",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Invalid idea id": {
										"value": {
											"errors": [
												"Please enter a valid idea id."
											]
										}
									},
									"Invalid vote": {
										"value": {
											"errors": [
												"Please enter a valid vote between 1 and 5."
											]
										}
									},
									"Empty comment": {
										"value": {
											"errors": [
												"Please enter a comment."
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/ideas/{ideaId}/members": {
			"get": {
				"summary": "Get idea current members",
				"operationId": "getCommunityIdeaMembers",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"$ref": "#/components/parameters/ideaId"
					}
				],
				"responses": {
					"200": {
						"description": "Members list",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Idea members": {
										"value": {
											"members": [
												{
													"user_id": 1,
													"user_name": "John Doe",
													"user_email": "john.doe@devscope.net",
													"role": "Member"
												}
											]
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/unexistent_idea"
					},
					"422": {
						"description": "Invalid fields",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Invalid idea id": {
										"value": {
											"errors": [
												"Please enter a valid idea id."
											]
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Add current user to idea members",
				"operationId": "addLoggedUserToMembers",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"$ref": "#/components/parameters/ideaId"
					}
				],
				"responses": {
					"200": {
						"description": "Member added to idea collaborators",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Member added": {
										"value": {
											"message": "User added to team members."
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Error on add",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Idea doesn't accept members (status)": {
										"value": {
											"error": "Idea is not open for members."
										}
									},
									"User is already member": {
										"value": {
											"error": "User is already in team."
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/unexistent_idea"
					},
					"422": {
						"description": "Invalid fields",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Invalid idea id": {
										"value": {
											"errors": [
												"Please enter a valid idea id."
											]
										}
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"summary": "Leave idea members",
				"operationId": "removeLoggedUserFromMembers",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"$ref": "#/components/parameters/ideaId"
					}
				],
				"responses": {
					"200": {
						"description": "Member left",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Member left": {
										"value": {
											"message": "User removed from team members."
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "User isn't a member",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Member error": {
										"value": {
											"error": "User is not in team."
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/unexistent_idea"
					},
					"422": {
						"description": "Invalid fields",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								},
								"examples": {
									"Invalid idea id": {
										"value": {
											"errors": [
												"Please enter a valid idea id."
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
	}
}